Add rule to pin to terraform-aws-modules module versions
We use the `terraform-aws-modules`[1] community modules a fair bit for
S3 buckets, Lambda functions and VPCs.

In recent versions they have started to ensure their major versions are
pinned to specific AWS provider versions.

We should always try to use the latest version of the module, but ensure
that it is pinned to the correct AWS provider, otherwise this can cause
issues.

This rule also ensures we use the rightmost operator for these modules,
to ensure we get the latest updates.

[1] https://registry.terraform.io/namespaces/terraform-aws-modules